requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.27
Jinja2==3.1.3
python-dotenv==1.0.1
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
matplotlib==3.8.3
seaborn==0.13.2
pytest==8.0.2
black==24.2.0
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pre-commit==3.6.0 